FIXED MAJOR BUG IN DYNAMICS ALGORITHM!
We were transforming the spatial acceleration a_hat_i of a rigid body attached to a floating joint from CoM frame to joint frame, in order to set the DoubleYoVariables qdd_*. However, the rest of the algorithm relies on a_hat_i being expressed in CoM frame! Now doing the same transformation on data that does not matter for the rest of the algorithm.

This fixes RobotTest.testSwitchingRootJoint(). Also added more robot tests.