Fixed hip roll bug (I hope)!
Please let me know if it ever happens again! I changed DRCDemo0Test.testDRCDemo0HipRollFlipOut so that it now actually fails if it were to flip out.

Here's what I fixed. We were using the previous solution for the spatial acceleration of the root joint from the momentum optimizer as the initial guess for the next tick. This somehow caused the optimizer to be unable to find a solution within the specified tolerance, which caused things to blow up eventually. I changed to using all zeros as the initial guess for the root joint spatial acceleration. The problem where the optimizer finds no solution within the required tolerance has not appeared since, and I am checking for this and throwing a RuntimeException if it does happen. Also let me know if you see this RuntimeException.