Added filter on the PelvisTrajectoryCommand based on selection matrix.
It used to be that sending a PelvisTrajectoryMessage would trigger the 3
pelvis manager even if the user only wants to control for instance the
pelvis height. This is troublesome especially for the XY manager, that
upon receiving a new message would start holding the pelvis position,
changing the controller behavior.